Restructure folder for better maintainablity and access
