+++ cs464-group-project/src/components/TeamComponents/Charts/CurrentTeamSchedule.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import '../../style/Team.css';
+import '../../../style/Team.css';
 
 import {
   Table,
