+++ cs464-group-project/src/components/TeamComponents/Components/TeamsSchedule.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
 import { useState, useEffect } from 'react';
-import { getUpcomingSchedule } from '../Api/ApiRequest';
-import { UpcomingSchedule } from '../charts/CurrentTeamSchedule';
-import '../../style/Team.css';
+import { getUpcomingSchedule } from '../../../Api/ApiRequest';
+import { UpcomingSchedule } from '../Charts/CurrentTeamSchedule';
+import '../../../style/Team.css';
 
 export function TeamsSchedule({ teamID }) {
   const location = useLocation();
