+++ cs464-group-project/src/pages/Players/IndividualPlayer.js
@@ -1,19 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import {
-  getPlayerDetails,
-  getTeamByName,
-} from '../../components/Api/ApiRequest';
+import { getPlayerDetails, getTeamByName } from '../../Api/ApiRequest';
 import '../../style/Player.css';
-import { PlayerFormerTeams } from '../../components/player/PlayerFormerTeams';
-import { PlayerHonours } from '../../components/player/PlayerHonours';
+import { PlayerFormerTeams } from '../../components/PlayerComponents/Components/PlayerFormerTeams';
+import { PlayerHonours } from '../../components/PlayerComponents/Components/PlayerHonours';
+import { PlayerFormerBadges } from '../../components/PlayerComponents/Components/PlayerFormerBadges';
 import {
   faInstagram,
   faXTwitter,
   faYoutube,
 } from '@fortawesome/free-brands-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { PlayerFormerBadges } from '../../components/player/PlayerFormerBadges';
 
 export function IndividualPlayer() {
   const { id } = useParams();
